Only call dwfl_attach_state with user-level CPU registers and stack
For samples with frame pointer or LBR call chains, we may not have
the CPU registers or stack information available. Under such
circumstances we would spam the command line with repeated warnings
of the form:

```
<pid> failed to attach state Couldn't find architecture of any ELF
```

Prevent that by checking whether the sample has register and stack
information available and only attach dwfl state then.

Note that symbol resolution and inline-frame resolution is independent
of the dwfl_attach_state call - that really is only required for stack
unwinding.

Based on a patch by vtoropov, thanks!

Fixes: https://github.com/kdab/hotspot/issues/178
Change-Id: I8b88ee5b97d3e9d3ded82a17c4487d01add6d5ca
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>