Allow zero sized symbols in the address cache
Apparently there are situations where the compiler generates entries
in the symbol table with zero size which are still referenced by some
callstacks. After commit fc12536 we ended up losing the symbol names
in such cases. This patch fixes it again and restores the symbol names
for such zero-size symbols.

Fixes: https://github.com/KDAB/hotspot/issues/252
Change-Id: Ib11521bc88f719811559d1e20f27c275eb8e5606