Use qCompress/qUncompress to reduce the size of the test input files
All files are now run through qCompress (manually) and then those
are tracked in git. When we run the test, the files are uncompressed
on the fly to let the rest of the code remain working as before.

The resulting folder is now only ~1.3MiB instead of ~5.2MiB before.
But obviously developers won't notice as the old data is still
included in our git history. But at least for new tests we can prevent
such pollutions and keep the gerrit bots happy.

Sadly we cannot resort to external tooling for the compression step,
so adding new files is a bit tedious but doable. And relying on
qCompress means we dont need any other new dependencies.

Change-Id: I902a6906f140eed2565df9637cb80cf464143b80
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>