Cache whether the perf-$pid.map file exists
When we parse perf data files with many processes, we would
continuously query whether the perf map file exists. QFile::exists
does not cache the result internally. In our scenario, it's enough
to just check the existence once and then be done with it. This
removes many stat syscalls for complex perf data files.

Change-Id: I452ea476cdcd7ac8ba3307c44f5fb70d17e73f10
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>