Partially revert "Optionally report debuginfod download progress"
This reverts the check for "libdwfl.h" and debuginfod added in commit
d432bc71926b1a855e54baabf5216b09b266b5fc and only the check. The code in
app/ that depended on this feature remains behind, pending a proper
detection of the elfutils content.

The problem is that elfutils 0.188 does have libdwfl.h and the symbol
that this check checked for, so the test passes. However, nothing
provides the CMake target elfutils::debuginfod: elfutils does not ship
with CMake files, CMake doesn't have Findelfutis.cmake, and Qt Creator's
Findelfutils.cmake doesn't provide this target.

Change-Id: Ib1d2fc7100134f7597cdfffd1749f527d965cf18
Reviewed-by: Milian Wolff <milian.wolff@kdab.com>