Don't duplicate the last inline frame in call stacks
Properly set addressLocation.parentLocationId after resolving the
inline chain instead of setting the functionLocation.parentLocationId.
With the previous code, we then ran into the conditionals below that
checked whether the parentLocationId was set - since it wasn't we
then added the last frame once more which is incorrect and lead to
confusing callstacks.

Change-Id: I09b25d623888521db7f2de9a174dd0acd9b271ea