Add basic support for PERF_FORMAT_LOST in PerfRecordSample parsing
Without this, we incorrectly interpreted the binary data stream
leading to totally bogus failures. Most notably, we ended up using
the id of the PERF_FORMAT_LOST as the number of callchain samples
which was really bad as that number is often very large, leading to
excessive memory consumption.

We do not actually read the lost values anywhere, this can come
in a follow-up patch next.

Fixes: https://github.com/KDAB/hotspot/issues/578
Change-Id: I3f73a5f02507fd31e21577585dd89679312782d9