improve file lookup time
On Debian/Ubuntu, there is a recursive symlink at /usr/bin/X11. Previously
we spend a lot of time traversing this loop until we hit MAX_PATH, which
is extremely wasteful.
Now we ensure that we only visit each canonical path once instead. In one
test, this makes a difference between more than 50s to find a handful of
libraries down to less than a second with this patch here.