fix: read base address from ELF headers and use it consistently
Some linkers such as lld and mold will create position-independent
binaries with p_offset != p_vaddr. This causes problems for some code in
hotspot which was sometimes assuming that they are the same (for example
by using addr - offset as the base address) and resulted in incorrect
backtraces. To fix this, make it so that we set the base address for
every readable binary by reading the ELF headers to check whether the
mapping fully contains a segment, and if so use the p_vaddr field to
calculate the base address. Subsequent code is adjusted to use the base
address instead of vaddr - offset if present.