fix: subtract 1 from call chain addresses other than the first one
For stack frames other than the first one, ip is a return address, so it
points to the instruction after the call instruction. By using the return
address as is, we were incorrectly attributing these call chain samples to
the instruction following the call. Therefore, subtract 1 from the return
address so that it gets correctly attributed to the call instruction.